+++ lib/widget/widget.dart
@@ -2,6 +2,8 @@ export 'niku.dart';
 
 export 'alert.dart';
 export 'button.dart';
+export 'checkbox.dart';
+export 'checkboxListTile.dart';
 export 'column.dart';
 export 'dropdownButton.dart';
 export 'gridView.dart';
@@ -12,6 +14,7 @@ export 'slider.dart';
 export 'selectableText.dart';
 export 'stack.dart';
 export 'switch.dart';
+export 'popupMenuButton.dart';
 export 'rangeSlider.dart';
 export 'text.dart';
 export 'textFormField.dart';